Core/SqlStatementModel/Unresolved/UnresolvedDummyRowTableInfo.cs
// This file is part of the re-linq project (relinq.codeplex.com)
// Copyright (c) rubicon IT GmbH, www.rubicon.eu
// 
// re-linq is free software; you can redistribute it and/or modify it under 
// the terms of the GNU Lesser General Public License as published by the 
// Free Software Foundation; either version 2.1 of the License, 
// or (at your option) any later version.
// 
// re-linq is distributed in the hope that it will be useful, 
// but WITHOUT ANY WARRANTY; without even the implied warranty of 
// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the 
// GNU Lesser General Public License for more details.
// 
// You should have received a copy of the GNU Lesser General Public License
// along with re-linq; if not, see http://www.gnu.org/licenses.
// 
using System;
using Remotion.Linq.SqlBackend.SqlStatementModel.Resolved;
using Remotion.Utilities;

namespace Remotion.Linq.SqlBackend.SqlStatementModel.Unresolved
{
  /// <summary>
  /// <see cref="UnresolvedDummyRowTableInfo"/> represents a table that returns exactly one row, with unspecified contents.
  /// </summary>
  public sealed class UnresolvedDummyRowTableInfo : ITableInfo
  {
    public static readonly UnresolvedDummyRowTableInfo Instance = new UnresolvedDummyRowTableInfo();

    private UnresolvedDummyRowTableInfo ()
    {
    }

    public Type ItemType
    {
      get { return typeof(object);  }
    }

    public ITableInfo Accept (ITableInfoVisitor visitor)
    {
      ArgumentUtility.CheckNotNull ("visitor", visitor);
      return visitor.VisitUnresolvedDummyRowTableInfo (this);
    }

    public IResolvedTableInfo GetResolvedTableInfo ()
    {
      throw new InvalidOperationException ("This table has not yet been resolved; call the resolution step first.");
    }

    public override string ToString ()
    {
      return "TABLE(one row)";
    }
  }
}